#65 - Refactor: 게시글 댓글 페이지 컴포넌트화
