+++ src/components/TopMenuBar/TopMenuBar.jsx
@@ -1,15 +1,28 @@
 import React from 'react';
+import { Link, useHistory } from 'react-router-dom';
 import styled from 'styled-components';
-import LeftArrow from '../../assets/icon/icon-arrow-left.png';
 import Button from '../Button/Button';
 
-export default function TopMenuBar({ saveButton }) {
+import LeftArrow from '../../assets/icon/icon-arrow-left.png';
+import MoreButton from '../Button/MoreButton';
+
+export default function TopMenuBar({
+  saveButton,
+  moreButton,
+  moreButtonSmall,
+  menuText,
+}) {
+  const history = useHistory();
+
   return (
     <Container>
-      <PreviousBtn>
+      <PreviousBtn onClick={() => history.goBack()}>
         <PrevioudBtnImg src={LeftArrow} />
       </PreviousBtn>
+      <MenuText>{menuText}</MenuText>
       {saveButton ? <Button buttonText="저장" size="medium-small" /> : null}
+      {moreButton ? <MoreButton size="large" /> : null}
+      {moreButtonSmall ? <MoreButton size="small" /> : null}
     </Container>
   );
 }
@@ -17,13 +30,24 @@ export default function TopMenuBar({ saveButton }) {
 const Container = styled.article`
   display: flex;
   justify-content: space-between;
+  align-items: center;
   border-bottom: 1px solid #dbdbdb;
   padding: 8px 20px;
 `;
 
-const PreviousBtn = styled.button``;
+const PreviousBtn = styled(Link)`
+  width: 22px;
+  border: none;
+`;
 
 const PrevioudBtnImg = styled.img`
   vertical-align: -4px;
   padding: 5px 0;
+`;
+
+const MenuText = styled.h1`
+  margin-left: 10px;
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 18px;
 `;
